feat: 셔틀 상세 마크업
