+++ src/components/shuttle-detail/ShuttleDetailPage.tsx
@@ -1,9 +1,6 @@
 import Footer from '@/components/footer/Footer';
-import ShuttleDemandStatus, {
-  SHUTTLE_DEMAND_STATUS_TYPE,
-} from './components/ShuttleDemandStatus';
+import ShuttleDemandStatus from './components/ShuttleDemandStatus';
 import ShuttleRouteVisualizer from '@/components/shuttle/shuttle-route-visualizer/ShuttleRouteVisualizer';
-
 import { SECTION } from '@/types/shuttle.types';
 import KakaoMap from './components/KakaoMap';
 import ShuttleImage from './components/ShuttleImage';
@@ -13,11 +10,8 @@ import NoticeSection, {
   NOTICE_TYPE,
 } from '@/components/notice-section/NoticeSection';
 import ShuttleForm from './components/ShuttleForm';
-import { BottomBarType } from './components/BottomBar';
 import Spacer from './components/Spacer';
-// import { BottomBarType } from './components/BottomBar';
 
-// 목킹용 임시 데이터
 const title = 'ATEEZ 2024 FANMEETING 〈ATINY’S VOYAGE FROM A TO Z〉';
 const artist = 'ATEEZ';
 const date = '2024. 07. 06. (토) ~ 2024. 07. 07. (일)';
@@ -29,7 +23,7 @@ const shuttle_date = '2024년 7월 6일 (토)';
 
 interface Props {
   shuttleId: string;
-  type: BottomBarType;
+  type: 'RESERVATION' | 'DEMAND';
 }
 const ShuttleDetailPage = ({ shuttleId, type }: Props) => {
   return (
@@ -52,31 +46,12 @@ const ShuttleDetailPage = ({ shuttleId, type }: Props) => {
       <div id="divider" className="my-16 h-[8px] bg-grey-50" />
       {type === 'RESERVATION' && (
         <ShuttleRouteVisualizer
-          object={[
-            { time: '2024-03-20 14:30:00', location: '청주터미널' },
-            {
-              time: '2024-03-20 14:40:00',
-              location: '청주대학교',
-              is_pickup: true,
-            },
-            { time: '2024-03-20 14:50:00', location: '장소3' },
-            { time: '2024-03-20 15:00:00', location: '장소4' },
-            {
-              time: '2024-03-20 15:10:00',
-              location: '장소5',
-              is_dropoff: true,
-            },
-            { time: '2024-03-20 15:20:00', location: '장소6' },
-          ]}
+          object={ROUTE_OBJECT}
           section={SECTION.SHUTTLE_DETAIL}
         />
       )}
       <ShuttleDemandStatus
-        type={
-          type === 'DEMAND'
-            ? SHUTTLE_DEMAND_STATUS_TYPE.DEMAND_SURVEY
-            : SHUTTLE_DEMAND_STATUS_TYPE.SELECT_SHUTTLE
-        }
+        type={type === 'DEMAND' ? 'DEMAND_SURVEY' : 'SELECT_SHUTTLE'}
         shuttle_date={shuttle_date}
         shuttle_location={shuttle_location}
       />
@@ -93,3 +68,20 @@ const ShuttleDetailPage = ({ shuttleId, type }: Props) => {
 };
 
 export default ShuttleDetailPage;
+
+const ROUTE_OBJECT = [
+  { time: '2024-03-20 14:30:00', location: '청주터미널' },
+  {
+    time: '2024-03-20 14:40:00',
+    location: '청주대학교',
+    is_pickup: true,
+  },
+  { time: '2024-03-20 14:50:00', location: '장소3' },
+  { time: '2024-03-20 15:00:00', location: '장소4' },
+  {
+    time: '2024-03-20 15:10:00',
+    location: '장소5',
+    is_dropoff: true,
+  },
+  { time: '2024-03-20 15:20:00', location: '장소6' },
+];
